requirements.txt
streamlit
requests
python-dotenv
pandas
langchain
langchain_groq
python-dotenv
